remove loading and offline segments, move functionality back to async-rendering
